+++ R/Kernel_loader.R
@@ -70,9 +70,13 @@
 #' @param one.lam Logical. When using formula input, this is a convenient way of
 #'   letting the function know to treat all variables as a single variable (i.e.
 #'   shared scale parameter). Defaults to \code{FALSE}.
+#' @param train.samp (Optional) A vector indicating which of the data points
+#'   should be used for training, and the remaining used for testing.
+#' @param test.samp (Optional) Similar to \code{train.samp}, but on test samples
+#'   instead.
 #'
-#' @return An \code{ipriorKernel} object which contains the relevant material
-#'   to be passed to the \code{iprior} function for model fitting.
+#' @return An \code{ipriorKernel} object which contains the relevant material to
+#'   be passed to the \code{iprior} function for model fitting.
 #'
 #' @seealso \link[=iprior]{iprior}
 #'
@@ -100,7 +104,7 @@ kernL <- function(
   y, ..., kernel = "linear", interactions = NULL, est.lambda = TRUE,
   est.hurst = FALSE, est.lengthscale = FALSE, est.offset = FALSE,
   est.psi = TRUE, fixed.hyp = NULL, lambda = 1, psi = 1, nystrom = FALSE,
-  nys.seed = NULL, model = list()
+  nys.seed = NULL, model = list(), train.samp, test.samp
 ) UseMethod("kernL")
 
 #' @export
@@ -109,13 +113,12 @@ kernL.default <- function(y, ..., kernel = "linear", interactions = NULL,
                            est.lengthscale = FALSE, est.offset = FALSE,
                            est.psi = TRUE, fixed.hyp = NULL, lambda = 1,
                            psi = 1, nystrom = FALSE, nys.seed = NULL,
-                           model = list()) {
+                           model = list(), train.samp, test.samp) {
   # Checks ---------------------------------------------------------------------
   if (is.list(model) & length(model) > 0) {
     stop("\'model\' option is deprecated. Use the arguments directly instead. See \'?kernL\' for details.", call. = FALSE)
   }
   kernel <- tolower(kernel)
-
   Xl <- list(...)
   # It is common to make the mistake and type kernels instead of kernel. This
   # corrects it.
@@ -124,23 +127,54 @@ kernL.default <- function(y, ..., kernel = "linear", interactions = NULL,
     kernel <- Xl[[Xl.kernel.mistake]]
     Xl[[Xl.kernel.mistake]] <- NULL
   }
+  # Check formula
   Xl.formula <- match("Xl.formula", names(Xl))
   formula.method <- FALSE
   if ("Xl.formula" %in% names(Xl)) {
     Xl <- Xl[[Xl.formula]]
     formula.method <- TRUE
   }
+  # Get names
   xname <- names(Xl)
   yname <- attr(y, "yname")
+  # Check for probit model
   if (is.factor(y)) {
-    probit <- TRUE
     tmp <- get_y_and_levels(y)
     y <- tmp$y
     y.levels <- tmp$levels
   } else {
-    probit <- FALSE
     y.levels <- NULL
   }
+  one.lam <- attr(Xl, "one.lam")
+
+  # Were training samples specified? -------------------------------------------
+  train.check <- FALSE
+  if (!missing(train.samp) | !missing(test.samp)) {
+    if (!missing(train.samp) & !missing(test.samp)) {
+      stop("Use either train.samp or test.samp only.", call. = FALSE)
+    }
+    if (missing(train.samp)) train.samp <- seq_along(y)[-test.samp]
+    if (all(train.samp %in% seq_along(y))) {
+      train.check <- TRUE
+      test.samp <- seq_along(y)[-train.samp]
+      y.test <- y[test.samp]
+      Xl.test <- lapply(Xl, function(x) {
+        if (is.matrix(x) | is.data.frame(x)) return(x[test.samp, , drop = FALSE])
+        else return(x[test.samp])
+      })
+      y <- y[train.samp]
+      Xl <- lapply(Xl, function(x) {
+        if (is.matrix(x) | is.data.frame(x)) return(x[train.samp, , drop = FALSE])
+        else return(x[train.samp])
+      })
+    } else {
+      warning("Training samples incorrectly specified.", call. = FALSE)
+    }
+    if (!is.null(one.lam)) attr(Xl, "one.lam") <- attr(Xl.test, "one.lam") <-
+        one.lam
+  }
+
+  # Get intercept --------------------------------------------------------------
   y <- scale(y, scale = FALSE)  # centre variables
   intercept <- attr(y, "scaled:center")
 
@@ -267,7 +301,7 @@ kernL.default <- function(y, ..., kernel = "linear", interactions = NULL,
     # Data
     y = y, Xl = Xl, Hl = Hl, intercept = intercept,
     # Model
-    kernels = kernels, which.pearson = which.pearson, probit = probit,
+    kernels = kernels, which.pearson = which.pearson,
     poly.deg = poly.deg, thetal = thetal, estl = estl,
     intr = intr, intr.3plus = intr.3plus, nystroml = nystroml,
     BlockBStuff = BlockBStuff,
@@ -276,9 +310,11 @@ kernL.default <- function(y, ..., kernel = "linear", interactions = NULL,
     xname = xname, yname = yname, formula = NULL, terms = NULL,
     y.levels = y.levels
   )
-
-  # Function call --------------------------------------------------------------
-  res$call <- fix_call_default(match.call(), "kernL")
+  if (isTRUE(train.check)) {
+    res$y.test <- y.test
+    res$Xl.test <- Xl.test
+  }
+  res$call <- fix_call_default(match.call(), "kernL")  # fix function call
 
   class(res) <- "ipriorKernel"
   res
@@ -291,67 +327,27 @@ kernL.formula <- function(formula, data, kernel = "linear", one.lam = FALSE,
                            est.lengthscale = FALSE, est.offset = FALSE,
                            est.psi = TRUE, fixed.hyp = NULL, lambda = 1,
                            psi = 1, nystrom = FALSE, nys.seed = NULL,
-                           model = list(), ...) {
-  mf <- model.frame(formula = formula, data = data)
-  tt <- terms(mf)
-  Terms <- delete.response(tt)
-  x <- model.frame(Terms, mf)
-  y <- model.response(mf)
-  yname <- names(attr(tt, "dataClasses"))[1]
-  xname <- names(x)
-  xnl <- length(xname)
-  x <- as.list(x)
-  attr(x, "terms") <- NULL
-  # attr(y, "yname") <- yname
-
-  # Interactions ---------------------------------------------------------------
-  interactions <- NULL
-  tmpo <- attr(tt, "order")
-  tmpf <- attr(tt, "factors")
-  tmpf2 <- as.matrix(tmpf[-1, tmpo == 2])  # this obtains 2nd order interactions
-  int2 <- apply(tmpf2, 2, function(x) which(x == 1))
-  if (any(tmpo == 2)) interactions <- int2
-  intr.3plus <- NULL
-  tmpf3 <- as.matrix(tmpf[-1, tmpo > 2])
-  int3 <- apply(tmpf3, 2, whereInt)
-  if (any(tmpo > 2)) intr.3plus <- int3
-  interactions <- list(intr = interactions, intr.3plus = intr.3plus)
-
-  # Deal with one.lam option ---------------------------------------------------
-  if (isTRUE(one.lam)) {
-    if (!all(sapply(interactions, is.null))) {
-      stop("Cannot use option one.lam = TRUE with interactions.", call. = FALSE)
-    }
-    if (length(x) == 1) {
-      message("Option one.lam = TRUE used with a single covariate anyway.")
-    }
-    attributes(x)$terms <- attributes(x)$names <- NULL
-    if (xnl <= 3) {
-      xname <- paste(xname, collapse = " + ")
-    } else {
-      xname <- paste(xname[1], "+ ... +", xname[xnl])
-    }
-    x <- list(matrix(unlist(x), ncol = length(x)))
-    names(x) <- xname
-  }
-
-  res <- kernL.default(y = y, Xl.formula = x, interactions = interactions,
+                           model = list(), train.samp, test.samp, ...) {
+  list2env(formula_to_xy(formula = formula, data = data, one.lam = one.lam),
+           envir = environment())
+  res <- kernL.default(y = y, Xl.formula = Xl, interactions = interactions,
                        kernel = kernel, est.lambda = est.lambda,
                        est.hurst = est.hurst,
                        est.lengthscale = est.lengthscale,
                        est.offset = est.offset, est.psi = est.psi,
                        fixed.hyp = fixed.hyp, lambda = lambda, psi = psi,
-                       nystrom = nystrom, nys.seed = nys.seed, model = model)
+                       nystrom = nystrom, nys.seed = nys.seed, model = model,
+                       train.samp = train.samp, test.samp = test.samp, ...)
   res$yname <- yname
   res$formula <- formula
   res$terms <- tt
-  res$call <- fix_call_formula(match.call(), "kernL")
+  res$call <- fix_call_formula(match.call(), "kernL")  # fix function call
 
   res
 }
 
 #' @export
-print.ipriorKernel <- function(x, units = "MB", standard = "SI", ...) {
+print.ipriorKernel <- function(x, units = "auto", standard = "SI", ...) {
   tmp <- expand_Hl_and_lambda(x$Hl, seq_along(x$Hl), x$intr, x$intr.3plus)
 
   # if (isTRUE(x$probit)) {
@@ -360,7 +356,12 @@ print.ipriorKernel <- function(x, units = "MB", standard = "SI", ...) {
   #   cat("Nystrom kernel approximation ()\n")
   # }
 
-  cat("Sample size:", x$n, "\n")
+  n <- x$n
+  if (is.ipriorKernel_cv(x)) {
+    n.test <- length(x$y.test)
+    n <- paste0(n + n.test, " (", n, " train + ", n.test, " test)")
+  }
+  cat("Sample size:", n, "\n")
   cat("No. of covariates:", length(x$Xl), "\n")
   # cat("No. of interactions:", x$no.int + x$no.int.3plus, "\n")
   cat("Object size: ")
@@ -392,12 +393,13 @@ print.ipriorKernel <- function(x, units = "MB", standard = "SI", ...) {
   if (is.nystrom(x)) {
     poss.method <- paste(poss.method, "(Nystrom)")
   }
-  if (is.iprobit(x)) {
+  if (is.categorical(x)) {
     poss.method <- c(poss.method, "iprobit (recommended)")
   }
   poss.method <- paste0(unique(poss.method), collapse = ", ")
   cat("Estimation methods available:\n")
   cat(poss.method)
+  cat("\n")
 }
 
 print_kern <- function(x, ...) {